Split out JSX templates
