+++ src/pages/_app.ts
@@ -1,0 +1,4 @@
+import '@styles/globals.css'
+import {App} from '@templates/App'
+
+export default App
