--- src/pages/_app.tsx
@@ -1,8 +1,0 @@
-import type {AppProps} from 'next/app'
-import '@styles/globals.css'
-
-const MyApp = ({Component, pageProps}: AppProps) => (
-	<Component {...pageProps} />
-)
-
-export default MyApp
