+++ src/pages/index.ts
@@ -1,0 +1,3 @@
+import {Home} from '@templates/Home'
+
+export default Home
