--- src/pages/index.tsx
@@ -1,36 +1,0 @@
-import type {NextPage} from 'next'
-import Head from 'next/head'
-import {Footer} from '@components/Footer'
-import styles from '@styles/Home.module.css'
-
-const Home: NextPage = ( ) => (
-	<div className={styles.container}>
-		<Head>
-			<title>Create Next App</title>
-			<meta
-				name='description'
-				content='Generated by create next app'
-			/>
-			<link
-				rel='icon'
-				href='/favicon.ico'
-			/>
-		</Head>
-
-		<main className={styles.main}>
-			<h1 className={styles.title}>
-				Welcome to{' '}
-				<a href='https://nextjs.org'>Next.js!</a>
-			</h1>
-
-			<p className={styles.description}>
-				Get started by editing{' '}
-				<code className={styles.code}>pages/index.tsx</code>
-			</p>
-		</main>
-
-		<Footer />
-	</div>
-)
-
-export default Home
