+++ src/templates/App/App.tsx
@@ -1,0 +1,27 @@
+import type {AppProps} from 'next/app'
+import Head from 'next/head'
+import {Footer} from '@components/Footer'
+import styles from '@styles/Home.module.css'
+
+const App = ({Component, pageProps}: AppProps) => (
+	<div className={styles.container}>
+		<Head>
+			<title>Create Next App</title>
+			<meta
+				name='description'
+				content='Generated by create next app'
+			/>
+			<link
+				rel='icon'
+				href='/favicon.ico'
+			/>
+		</Head>
+
+		{/* Component injection point is here! */}
+		<Component {...pageProps} />
+
+		<Footer />
+	</div>
+)
+
+export {App}
